feat: replace generate proto tool into buf
