+++ packages/bindings/src/types.rs
@@ -1,6 +1,6 @@
 //! Contains basic types of a cosmos sdk based chain.
 
 pub use desmos_std::proto::cosmos::base::query::v1beta1::PageRequest;
+pub use desmos_std::proto::cosmos::crypto::{ed25519, secp256k1, secp256r1};
 pub use desmos_std::proto::ibc::core::client::v1::Height;
-pub use desmos_std::public_keys::{Ed25519PublicKey, Secp256k1PublicKey, Secp256r1PublicKey};
 pub use desmos_std::shim::Any;
