+++ packages/proto-build/Cargo.toml
@@ -16,9 +16,4 @@ prost-types = "0.11.1"
 regex = "1"
 syn = {version = "1.0.103", features = ["full", "parsing", "extra-traits"]}
 tonic = "0.8.2"
-tonic-build = "0.8.2"
 walkdir = "2"
-proc-macro2 = "1.0.18"
-pbjson = "0.5.1"
-pbjson-types = "0.5.1"
-pbjson-build = "0.5.1"
