+++ packages/proto-build/buf.gen.yaml
@@ -1,0 +1,8 @@
+version: v1
+plugins:
+  - plugin: buf.build/community/neoeinstein-prost
+    out: .
+    opt:
+      - extern_path=.google.protobuf.Timestamp=crate::shim::Timestamp
+      - extern_path=.google.protobuf.Duration=crate::shim::Duration
+      - extern_path=.google.protobuf.Any=crate::shim::Any
