+++ packages/proto-build/buf.lock
@@ -1,0 +1,19 @@
+# Generated by buf. DO NOT EDIT.
+version: v1
+deps:
+  - remote: buf.build
+    owner: cosmos
+    repository: cosmos-proto
+    commit: 1935555c206d4afb9e94615dfd0fad31
+  - remote: buf.build
+    owner: cosmos
+    repository: cosmos-sdk
+    commit: 302a6a6ac7cf478698eced661049a290
+  - remote: buf.build
+    owner: cosmos
+    repository: gogo-proto
+    commit: 34d970b699f84aa382f3c29773a60836
+  - remote: buf.build
+    owner: googleapis
+    repository: googleapis
+    commit: 5ae7f88519b04fe1965da0f8a375a088
