+++ packages/proto-build/src/code_generator.rs
@@ -1,27 +1,37 @@
 use std::fs::{create_dir_all, remove_dir_all};
 use std::path::{Path, PathBuf};
 use std::process::Command;
-use std::{env, fs};
+use std::{env, fs, io};
 
 use log::info;
 use prost::Message;
 use prost_types::FileDescriptorSet;
-use regex::Regex;
 use walkdir::WalkDir;
 
 use crate::{mod_gen, transform};
 
-const DESCRIPTOR_FILE: &str = "descriptor.bin";
+const UNSUPPORTED_MODULE: &[&str] = &[
+    // currently unsupported due to dependency on tendermint-proto
+    "cosmos.base.abci",
+    "cosmos.base.kv",
+    "cosmos.base.reflection",
+    "cosmos.base.store",
+    "cosmos.base.snapshots",
+    "cosmos.base.tendermint",
+];
 
+#[derive(Clone)]
 pub struct CosmosProject {
     pub name: String,
     pub version: String,
     pub project_dir: String,
+
+    /// determines which modules to include from the project
+    pub include_mods: Vec<String>,
 }
 
 pub struct CodeGenerator {
     project: CosmosProject,
-    tonic_build_config: tonic_build::Builder,
     root: PathBuf,
     out_dir: PathBuf,
     tmp_build_dir: PathBuf,
@@ -36,16 +46,8 @@ impl CodeGenerator {
         project: CosmosProject,
         deps: Vec<CosmosProject>,
     ) -> Self {
-        let tonic_build_config = tonic_build::configure()
-            .build_client(false)
-            .build_server(false)
-            .extern_path(".google.protobuf.Timestamp", "crate::shim::Timestamp")
-            .extern_path(".google.protobuf.Duration", "crate::shim::Duration")
-            .extern_path(".google.protobuf.Any", "crate::shim::Any");
-
         Self {
             project,
-            tonic_build_config,
             root: PathBuf::from(env!("CARGO_MANIFEST_DIR")),
             out_dir,
             tmp_build_dir,
@@ -62,6 +64,7 @@ impl CodeGenerator {
             self.project.name
         );
 
+        self.exclude_unsupported_module();
         self.transform();
         self.generate_mod_file();
         self.fmt();
@@ -82,6 +85,26 @@ impl CodeGenerator {
             &self.project.version,
             &self.tmp_namespaced_dir(),
         );
+    }
+
+    fn exclude_unsupported_module(&self) {
+        for entry in WalkDir::new(self.tmp_namespaced_dir()) {
+            let entry = entry.unwrap();
+            if entry.file_type().is_file() {
+                let filename = entry
+                    .file_name()
+                    .to_os_string()
+                    .to_str()
+                    .unwrap()
+                    .to_string();
+                if UNSUPPORTED_MODULE
+                    .iter()
+                    .any(|module| filename.contains(module))
+                {
+                    fs::remove_file(entry.path()).unwrap();
+                }
+            }
+        }
     }
 
     fn generate_mod_file(&self) {
@@ -120,44 +143,86 @@ impl CodeGenerator {
     }
 
     fn compile_proto(&self) {
-        let include_paths = ["proto", "third_party/proto"];
+        let buf_gen_template = self.root.join("buf.gen.yaml");
 
-        let deps_dirs = self
-            .deps
-            .iter()
-            .map(|dep| self.root.join(&dep.project_dir))
-            .collect();
-        let project_dir = self.root.join(&self.project.project_dir);
-
-        let proto_includes_path = vec![deps_dirs, vec![project_dir.clone()]].concat();
-        let proto_includes_paths = proto_includes_path
-            .iter()
-            .flat_map(|dir| include_paths.iter().map(|path| dir.join(path)));
-
-        // List available paths for dependencies
-        let includes: Vec<PathBuf> = proto_includes_paths.map(PathBuf::from).collect();
-
-        let proto_paths = fs::read_dir(project_dir.join(format!("proto/{}", self.project.name)))
-            .unwrap()
-            .map(|d| d.unwrap().path().to_string_lossy().to_string())
-            .collect::<Vec<String>>();
-
-        // List available proto files
-        let mut protos: Vec<PathBuf> = vec![];
-        collect_protos(&proto_paths, &mut protos);
+        let all_related_projects = vec![self.deps.clone(), vec![self.project.clone()]].concat();
 
         info!(
             "🧪 [{}] Compiling types from protobuf definitions...",
             self.project.name
         );
 
-        let descriptor_file = self.tmp_namespaced_dir().join(DESCRIPTOR_FILE);
-        self.tonic_build_config
-            .clone()
-            .out_dir(self.tmp_namespaced_dir())
-            .file_descriptor_set_path(&descriptor_file)
-            .compile(&protos, &includes)
-            .unwrap();
+        // Compile proto files for each file in `protos` variable
+        // `buf generate —template {<buf_gen_template} <proto_file>`
+        for project in all_related_projects {
+            let proto_path = &self.root.join(&project.project_dir).join("proto");
+
+            let buf_root = WalkDir::new(&self.root.join(proto_path))
+                .into_iter()
+                .filter_map(|e| e.ok())
+                .find(|e| {
+                    e.file_name()
+                        .to_str()
+                        .map(|s| s == "buf.yaml" || s == "buf.yml")
+                        .unwrap_or(false)
+                })
+                .map(|e| e.path().parent().unwrap().to_path_buf())
+                .unwrap();
+
+            
+
+            let mut cmd = Command::new("buf");
+            cmd.arg("generate")
+                .arg(buf_root.to_string_lossy().to_string())
+                .arg("--template")
+                .arg(buf_gen_template.to_string_lossy().to_string())
+                .arg("--output")
+                .arg(self.tmp_namespaced_dir().to_string_lossy().to_string());
+
+            if !project.include_mods.is_empty() {
+                for include_mod in project.include_mods.clone() {
+                    cmd.arg("--path")
+                        .arg(proto_path.join(project.name.clone()).join(include_mod));
+                }
+            }
+
+            let exit_status = cmd.spawn().unwrap().wait().unwrap();
+
+            if !exit_status.success() {
+                panic!(
+                    "unable to generate with: {:?}",
+                    cmd.get_args().collect::<Vec<_>>()
+                );
+            }
+
+            let descriptor_file = self
+                .tmp_namespaced_dir()
+                .join(format!("descriptor_{}.bin", project.name));
+
+            // generate descriptor file with `buf build buf.yaml --as-file-descriptor-set -o {descriptor_file}`
+            let mut cmd = Command::new("buf");
+            cmd.arg("build")
+                .arg(buf_root.to_string_lossy().to_string())
+                .arg("--as-file-descriptor-set")
+                .arg("-o")
+                .arg(descriptor_file.to_string_lossy().to_string());
+
+            if !project.include_mods.is_empty() {
+                for include_mod in project.include_mods {
+                    cmd.arg("--path")
+                        .arg(proto_path.join(project.name.clone()).join(include_mod));
+                }
+            }
+
+            let exit_status = cmd.spawn().unwrap().wait().unwrap();
+
+            if !exit_status.success() {
+                panic!(
+                    "unable to build with: {:?}",
+                    cmd.get_args().collect::<Vec<_>>()
+                );
+            }
+        }
 
         info!(
             "✨  [{}] Types from protobuf definitions is compiled successfully!",
@@ -166,13 +231,37 @@ impl CodeGenerator {
     }
 
     pub fn file_descriptor_set(&self) -> FileDescriptorSet {
-        let descriptor_file = self.tmp_namespaced_dir().join(DESCRIPTOR_FILE);
-        let descriptor_bytes = &fs::read(descriptor_file).unwrap()[..];
+        // list all files in self.tmp_namespaced_dir()
+        let files = fs::read_dir(self.tmp_namespaced_dir())
+            .unwrap()
+            .map(|res| res.map(|e| e.path()))
+            .collect::<Result<Vec<_>, io::Error>>()
+            .unwrap();
 
-        FileDescriptorSet::decode(descriptor_bytes).unwrap()
+        // filter only files that match "descriptor_*.bin"
+        let descriptor_files = files
+            .iter()
+            .filter(|f| {
+                f.file_name()
+                    .unwrap()
+                    .to_str()
+                    .unwrap()
+                    .starts_with("descriptor_")
+            })
+            .collect::<Vec<_>>();
+
+        // read all files and merge them into one FileDescriptorSet
+        let mut file_descriptor_set = FileDescriptorSet { file: vec![] };
+        for descriptor_file in descriptor_files {
+            let descriptor_bytes = &fs::read(descriptor_file).unwrap()[..];
+            let mut file_descriptor_set_tmp = FileDescriptorSet::decode(descriptor_bytes).unwrap();
+            file_descriptor_set
+                .file
+                .append(&mut file_descriptor_set_tmp.file);
+        }
+
+        file_descriptor_set
     }
-
-    // TODO: create config tonic
 
     fn tmp_namespaced_dir(&self) -> PathBuf {
         self.tmp_build_dir.join(&self.project.name)
@@ -182,29 +271,6 @@ impl CodeGenerator {
 fn output_version_file(project_name: &str, versions: &str, out_dir: &Path) {
     let path = out_dir.join(format!("{}_COMMIT", project_name.to_uppercase()));
     fs::write(path, versions).unwrap();
-}
-
-/// collect_protos walks every path in `proto_paths` and recursively locates all .proto
-/// files in each path's subdirectories, adding the full path of each file to `protos`
-///
-/// Any errors encountered will cause failure for the path provided to WalkDir::new()
-fn collect_protos(proto_paths: &[String], protos: &mut Vec<PathBuf>) {
-    for proto_path in proto_paths {
-        protos.append(
-            &mut WalkDir::new(proto_path)
-                .into_iter()
-                .filter_map(|e| e.ok())
-                .filter(|e| {
-                    e.file_type().is_file()
-                        && e.file_name() != "genesis.proto" // ignore all modules genesis.proto
-                        && !Regex::new("/client/cli.proto$").unwrap().is_match(e.path().to_str().unwrap_or_default()) // ignore */client/cli.proto
-                        && e.path().extension().is_some()
-                        && e.path().extension().unwrap() == "proto"
-                })
-                .map(|e| e.into_path())
-                .collect(),
-        );
-    }
 }
 
 fn find_cargo_toml(path: &Path) -> PathBuf {
