+++ packages/std/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod proto;
-pub mod public_keys;
 mod serde;
 pub mod shim;
