+++ packages/std/src/proto/cosmos/base/query/v1beta1.rs
@@ -15,7 +15,7 @@
     serde::Deserialize,
     std_derive::CosmwasmExt,
 )]
-#[proto_message(type_url = "/cosmos.base.query.v1beta1.PageRequest")]
+#[proto_message(type_url = "/cosmos.base.query.v1beta1.PageRequestPageRequest")]
 #[serde(rename_all = "snake_case")]
 pub struct PageRequest {
     /// key is a value returned in PageResponse.next_key to begin
@@ -73,11 +73,12 @@ pub struct PageRequest {
     serde::Deserialize,
     std_derive::CosmwasmExt,
 )]
-#[proto_message(type_url = "/cosmos.base.query.v1beta1.PageResponse")]
+#[proto_message(type_url = "/cosmos.base.query.v1beta1.PageResponsePageResponse")]
 #[serde(rename_all = "snake_case")]
 pub struct PageResponse {
     /// next_key is the key to be passed to PageRequest.key to
-    /// query the next page most efficiently
+    /// query the next page most efficiently. It will be empty if
+    /// there are no more results.
     #[prost(bytes = "vec", tag = "1")]
     #[serde(
         serialize_with = "crate::serde::as_base64::serialize",
