+++ packages/std/src/proto/cosmos/crypto/mod.rs
@@ -1,1 +1,4 @@
+pub mod ed25519;
 pub mod multisig;
+pub mod secp256k1;
+pub mod secp256r1;
