+++ packages/std/src/proto/cosmos/crypto/multisig/mod.rs
@@ -1,1 +1,22 @@
 pub mod v1beta1;
+/// LegacyAminoPubKey specifies a public key type
+/// which nests multiple public keys and a threshold,
+/// it uses legacy amino address rules.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(
+    Clone,
+    PartialEq,
+    ::prost::Message,
+    schemars::JsonSchema,
+    serde::Serialize,
+    serde::Deserialize,
+    std_derive::CosmwasmExt,
+)]
+#[proto_message(type_url = "/cosmos.crypto.multisig.LegacyAminoPubKey")]
+#[serde(rename_all = "snake_case")]
+pub struct LegacyAminoPubKey {
+    #[prost(uint32, tag = "1")]
+    pub threshold: u32,
+    #[prost(message, repeated, tag = "2")]
+    pub public_keys: ::prost::alloc::vec::Vec<crate::shim::Any>,
+}
