+++ packages/std/src/proto/cosmos/crypto/secp256k1.rs
@@ -1,0 +1,46 @@
+/// PubKey defines a secp256k1 public key
+/// Key is the compressed form of the pubkey. The first byte depends is a 0x02 byte
+/// if the y-coordinate is the lexicographically largest of the two associated with
+/// the x-coordinate. Otherwise the first byte is a 0x03.
+/// This prefix is followed with the x-coordinate.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(
+    Clone,
+    PartialEq,
+    ::prost::Message,
+    schemars::JsonSchema,
+    serde::Serialize,
+    serde::Deserialize,
+    std_derive::CosmwasmExt,
+)]
+#[proto_message(type_url = "/cosmos.crypto.secp256k1.PubKey")]
+#[serde(rename_all = "snake_case")]
+pub struct PubKey {
+    #[prost(bytes = "vec", tag = "1")]
+    #[serde(
+        serialize_with = "crate::serde::as_base64::serialize",
+        deserialize_with = "crate::serde::as_base64::deserialize"
+    )]
+    pub key: ::prost::alloc::vec::Vec<u8>,
+}
+/// PrivKey defines a secp256k1 private key.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(
+    Clone,
+    PartialEq,
+    ::prost::Message,
+    schemars::JsonSchema,
+    serde::Serialize,
+    serde::Deserialize,
+    std_derive::CosmwasmExt,
+)]
+#[proto_message(type_url = "/cosmos.crypto.secp256k1.PrivKey")]
+#[serde(rename_all = "snake_case")]
+pub struct PrivKey {
+    #[prost(bytes = "vec", tag = "1")]
+    #[serde(
+        serialize_with = "crate::serde::as_base64::serialize",
+        deserialize_with = "crate::serde::as_base64::deserialize"
+    )]
+    pub key: ::prost::alloc::vec::Vec<u8>,
+}
