+++ packages/std/src/proto/cosmos/tx/signing/v1beta1.rs
@@ -9,7 +9,7 @@
     serde::Deserialize,
     std_derive::CosmwasmExt,
 )]
-#[proto_message(type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptors")]
+#[proto_message(type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptorsSignatureDescriptors")]
 #[serde(rename_all = "snake_case")]
 pub struct SignatureDescriptors {
     /// signatures are the signature descriptors
@@ -30,7 +30,7 @@ pub struct SignatureDescriptors {
     serde::Deserialize,
     std_derive::CosmwasmExt,
 )]
-#[proto_message(type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptor")]
+#[proto_message(type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptorSignatureDescriptor")]
 #[serde(rename_all = "snake_case")]
 pub struct SignatureDescriptor {
     /// public_key is the public key of the signer
@@ -61,7 +61,9 @@ pub mod signature_descriptor {
         serde::Deserialize,
         std_derive::CosmwasmExt,
     )]
-    #[proto_message(type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptor.Data")]
+    #[proto_message(
+        type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptor.DataSignatureDescriptor.Data"
+    )]
     #[serde(rename_all = "snake_case")]
     pub struct Data {
         /// sum is the oneof that specifies whether this represents single or multi-signature data
@@ -82,7 +84,9 @@ pub mod signature_descriptor {
             serde::Deserialize,
             std_derive::CosmwasmExt,
         )]
-        #[proto_message(type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptor.Data.Single")]
+        #[proto_message(
+            type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptor.Data.SingleSignatureDescriptor.Data.Single"
+        )]
         #[serde(rename_all = "snake_case")]
         pub struct Single {
             /// mode is the signing mode of the single signer
@@ -111,7 +115,9 @@ pub mod signature_descriptor {
             serde::Deserialize,
             std_derive::CosmwasmExt,
         )]
-        #[proto_message(type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptor.Data.Multi")]
+        #[proto_message(
+            type_url = "/cosmos.tx.signing.v1beta1.SignatureDescriptor.Data.MultiSignatureDescriptor.Data.Multi"
+        )]
         #[serde(rename_all = "snake_case")]
         pub struct Multi {
             /// bitarray specifies which keys within the multisig are signing
@@ -145,23 +151,37 @@ pub mod signature_descriptor {
     }
 }
 /// SignMode represents a signing mode with its own security guarantees.
+///
+/// This enum should be considered a registry of all known sign modes
+/// in the Cosmos ecosystem. Apps are not expected to support all known
+/// sign modes. Apps that would like to support custom  sign modes are
+/// encouraged to open a small PR against this file to add a new case
+/// to this SignMode enum describing their sign mode so that different
+/// apps have a consistent version of this enum.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 #[derive(strum_macros::FromRepr, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum SignMode {
     /// SIGN_MODE_UNSPECIFIED specifies an unknown signing mode and will be
-    /// rejected
+    /// rejected.
     Unspecified = 0,
     /// SIGN_MODE_DIRECT specifies a signing mode which uses SignDoc and is
-    /// verified with raw bytes from Tx
+    /// verified with raw bytes from Tx.
     Direct = 1,
     /// SIGN_MODE_TEXTUAL is a future signing mode that will verify some
     /// human-readable textual representation on top of the binary representation
-    /// from SIGN_MODE_DIRECT
+    /// from SIGN_MODE_DIRECT. It is currently not supported.
     Textual = 2,
+    /// SIGN_MODE_DIRECT_AUX specifies a signing mode which uses
+    /// SignDocDirectAux. As opposed to SIGN_MODE_DIRECT, this sign mode does not
+    /// require signers signing over other signers' `signer_info`. It also allows
+    /// for adding Tips in transactions.
+    ///
+    /// Since: cosmos-sdk 0.46
+    DirectAux = 3,
     /// SIGN_MODE_LEGACY_AMINO_JSON is a backwards compatibility mode which uses
-    /// Amino JSON and will be removed in the future
+    /// Amino JSON and will be removed in the future.
     LegacyAminoJson = 127,
     /// SIGN_MODE_EIP_191 specifies the sign mode for EIP 191 signing on the Cosmos
     /// SDK. Ref: <https://eips.ethereum.org/EIPS/eip-191>
@@ -185,6 +205,7 @@ impl SignMode {
             SignMode::Unspecified => "SIGN_MODE_UNSPECIFIED",
             SignMode::Direct => "SIGN_MODE_DIRECT",
             SignMode::Textual => "SIGN_MODE_TEXTUAL",
+            SignMode::DirectAux => "SIGN_MODE_DIRECT_AUX",
             SignMode::LegacyAminoJson => "SIGN_MODE_LEGACY_AMINO_JSON",
             SignMode::Eip191 => "SIGN_MODE_EIP_191",
         }
@@ -195,6 +216,7 @@ impl SignMode {
             "SIGN_MODE_UNSPECIFIED" => Some(Self::Unspecified),
             "SIGN_MODE_DIRECT" => Some(Self::Direct),
             "SIGN_MODE_TEXTUAL" => Some(Self::Textual),
+            "SIGN_MODE_DIRECT_AUX" => Some(Self::DirectAux),
             "SIGN_MODE_LEGACY_AMINO_JSON" => Some(Self::LegacyAminoJson),
             "SIGN_MODE_EIP_191" => Some(Self::Eip191),
             _ => None,
