--- packages/std/src/proto/desmos/fees/mod.rs
@@ -1,1 +1,0 @@
-pub mod v1;
