+++ packages/std/src/proto/desmos/mod.rs
@@ -1,8 +1,6 @@
-pub mod fees;
 pub mod posts;
 pub mod profiles;
 pub mod reactions;
 pub mod relationships;
 pub mod reports;
 pub mod subspaces;
-pub mod supply;
