--- packages/std/src/proto/desmos/supply/mod.rs
@@ -1,1 +1,0 @@
-pub mod v1;
